fix(cxl-ui): unify media queries
As per the discussion here:
https://cxlworld.slack.com/archives/C01JABH8AHX/p1693292365293889 there
is a discrepency in our usage of media queries where we toggle the wide
layout of the site, before our components toggle there phone media
state. Because of the complex logic in our navigation component, it is
also necessary to reload the page when the media query match state
changes.